Reusable checklist model
