+++ test/routes/checkBeforeSending.route.test.js
@@ -12,7 +12,7 @@ const server = require('../../server')
 const Application = require('../../src/persistence/entities/application.entity')
 const BaseTaskList = require('../../src/models/taskList/base.taskList')
 const BaseCheck = require('../../src/models/checkList/base.check')
-const CheckBeforeSendingController = require('../../src/controllers/checkBeforeSending.controller')
+const CheckList = require('../../src/models/checkList/checkList')
 const CookieService = require('../../src/services/cookie.service')
 const RecoveryService = require('../../src/services/recovery.service')
 const UploadService = require('../../src/services/upload.service')
@@ -95,7 +95,7 @@ lab.beforeEach(() => {
   sandbox.stub(Application.prototype, 'save').value(() => undefined)
   sandbox.stub(Application.prototype, 'isSubmitted').value(() => false)
   sandbox.stub(UploadService, 'upload').value(() => undefined)
-  sandbox.stub(CheckBeforeSendingController.prototype, 'Checks').get(() => [ValidCheck, InvalidCheck])
+  sandbox.stub(CheckList.prototype, 'Checks').get(() => [ValidCheck, InvalidCheck])
   sandbox.stub(BaseTaskList, 'getTaskListClass').value(() => TaskList)
   sandbox.stub(BaseTaskList, 'buildTaskList').value(() => new TaskList())
   sandbox.stub(BaseTaskList, 'isComplete').value(() => true)
